ree_util: Extract ReadWriteValue from vector_run_end_encode.cc to a header
