+++ cpp/src/arrow/compute/kernels/vector_run_end_encode.cc
@@ -20,6 +20,7 @@
 #include "arrow/compute/api_vector.h"
 #include "arrow/compute/kernel.h"
 #include "arrow/compute/kernels/common_internal.h"
+#include "arrow/compute/kernels/ree_util_internal.h"
 #include "arrow/type_traits.h"
 #include "arrow/util/checked_cast.h"
 #include "arrow/util/ree_util.h"
@@ -27,269 +28,6 @@
 namespace arrow {
 namespace compute {
 namespace internal {
-
-template <typename ArrowType, bool has_validity_buffer, typename Enable = void>
-struct ReadWriteValueImpl {};
-
-// Numeric and primitive C-compatible types
-template <typename ArrowType, bool has_validity_buffer>
-class ReadWriteValueImpl<ArrowType, has_validity_buffer,
-                         enable_if_has_c_type<ArrowType>> {
- public:
-  using ValueRepr = typename ArrowType::c_type;
-
- private:
-  const uint8_t* input_validity_;
-  const uint8_t* input_values_;
-
-  // Needed only by the writing functions
-  uint8_t* output_validity_;
-  uint8_t* output_values_;
-
- public:
-  explicit ReadWriteValueImpl(const ArraySpan& input_values_array,
-                              ArrayData* output_values_array_data)
-      : input_validity_(has_validity_buffer ? input_values_array.buffers[0].data
-                                            : NULLPTR),
-        input_values_(input_values_array.buffers[1].data),
-        output_validity_((has_validity_buffer && output_values_array_data)
-                             ? output_values_array_data->buffers[0]->mutable_data()
-                             : NULLPTR),
-        output_values_(output_values_array_data
-                           ? output_values_array_data->buffers[1]->mutable_data()
-                           : NULLPTR) {}
-
-  [[nodiscard]] bool ReadValue(ValueRepr* out, int64_t read_offset) const {
-    bool valid = true;
-    if constexpr (has_validity_buffer) {
-      valid = bit_util::GetBit(input_validity_, read_offset);
-    }
-    if constexpr (std::is_same_v<ArrowType, BooleanType>) {
-      *out = bit_util::GetBit(input_values_, read_offset);
-    } else {
-      *out = (reinterpret_cast<const ValueRepr*>(input_values_))[read_offset];
-    }
-    return valid;
-  }
-
-  /// \brief Ensure padding is zeroed in validity bitmap.
-  void ZeroValidityPadding(int64_t length) const {
-    DCHECK(output_values_);
-    if constexpr (has_validity_buffer) {
-      DCHECK(output_validity_);
-      const int64_t validity_buffer_size = bit_util::BytesForBits(length);
-      output_validity_[validity_buffer_size - 1] = 0;
-    }
-  }
-
-  void WriteValue(int64_t write_offset, bool valid, ValueRepr value) const {
-    if constexpr (has_validity_buffer) {
-      bit_util::SetBitTo(output_validity_, write_offset, valid);
-    }
-    if (valid) {
-      if constexpr (std::is_same_v<ArrowType, BooleanType>) {
-        bit_util::SetBitTo(output_values_, write_offset, value);
-      } else {
-        (reinterpret_cast<ValueRepr*>(output_values_))[write_offset] = value;
-      }
-    }
-  }
-
-  void WriteRun(int64_t write_offset, int64_t run_length, bool valid,
-                ValueRepr value) const {
-    if constexpr (has_validity_buffer) {
-      bit_util::SetBitsTo(output_validity_, write_offset, run_length, valid);
-    }
-    if (valid) {
-      if constexpr (std::is_same_v<ArrowType, BooleanType>) {
-        bit_util::SetBitsTo(reinterpret_cast<uint8_t*>(output_values_), write_offset,
-                            run_length, value);
-      } else {
-        auto* output_values_c = reinterpret_cast<ValueRepr*>(output_values_);
-        std::fill(output_values_c + write_offset,
-                  output_values_c + write_offset + run_length, value);
-      }
-    }
-  }
-
-  bool Compare(ValueRepr lhs, ValueRepr rhs) const { return lhs == rhs; }
-};
-
-// FixedSizeBinary, Decimal128
-template <typename ArrowType, bool has_validity_buffer>
-class ReadWriteValueImpl<ArrowType, has_validity_buffer,
-                         enable_if_fixed_size_binary<ArrowType>> {
- public:
-  // Every value is represented as a pointer to byte_width_ bytes
-  using ValueRepr = uint8_t const*;
-
- private:
-  const uint8_t* input_validity_;
-  const uint8_t* input_values_;
-
-  // Needed only by the writing functions
-  uint8_t* output_validity_;
-  uint8_t* output_values_;
-
-  const size_t byte_width_;
-
- public:
-  ReadWriteValueImpl(const ArraySpan& input_values_array,
-                     ArrayData* output_values_array_data)
-      : input_validity_(has_validity_buffer ? input_values_array.buffers[0].data
-                                            : NULLPTR),
-        input_values_(input_values_array.buffers[1].data),
-        output_validity_((has_validity_buffer && output_values_array_data)
-                             ? output_values_array_data->buffers[0]->mutable_data()
-                             : NULLPTR),
-        output_values_(output_values_array_data
-                           ? output_values_array_data->buffers[1]->mutable_data()
-                           : NULLPTR),
-        byte_width_(input_values_array.type->byte_width()) {}
-
-  [[nodiscard]] bool ReadValue(ValueRepr* out, int64_t read_offset) const {
-    bool valid = true;
-    if constexpr (has_validity_buffer) {
-      valid = bit_util::GetBit(input_validity_, read_offset);
-    }
-    *out = input_values_ + (read_offset * byte_width_);
-    return valid;
-  }
-
-  /// \brief Ensure padding is zeroed in validity bitmap.
-  void ZeroValidityPadding(int64_t length) const {
-    DCHECK(output_values_);
-    if constexpr (has_validity_buffer) {
-      DCHECK(output_validity_);
-      const int64_t validity_buffer_size = bit_util::BytesForBits(length);
-      output_validity_[validity_buffer_size - 1] = 0;
-    }
-  }
-
-  void WriteValue(int64_t write_offset, bool valid, ValueRepr value) const {
-    if constexpr (has_validity_buffer) {
-      bit_util::SetBitTo(output_validity_, write_offset, valid);
-    }
-    if (valid) {
-      memcpy(output_values_ + (write_offset * byte_width_), value, byte_width_);
-    }
-  }
-
-  void WriteRun(int64_t write_offset, int64_t run_length, bool valid,
-                ValueRepr value) const {
-    if constexpr (has_validity_buffer) {
-      bit_util::SetBitsTo(output_validity_, write_offset, run_length, valid);
-    }
-    if (valid) {
-      uint8_t* ptr = output_values_ + (write_offset * byte_width_);
-      for (int64_t i = 0; i < run_length; ++i) {
-        memcpy(ptr, value, byte_width_);
-        ptr += byte_width_;
-      }
-    }
-  }
-
-  bool Compare(ValueRepr lhs, ValueRepr rhs) const {
-    return memcmp(lhs, rhs, byte_width_) == 0;
-  }
-};
-
-// Binary, String...
-template <typename ArrowType, bool has_validity_buffer>
-class ReadWriteValueImpl<ArrowType, has_validity_buffer,
-                         enable_if_base_binary<ArrowType>> {
- public:
-  using ValueRepr = std::string_view;
-  using offset_type = typename ArrowType::offset_type;
-
- private:
-  const uint8_t* input_validity_;
-  const offset_type* input_offsets_;
-  const uint8_t* input_values_;
-
-  // Needed only by the writing functions
-  uint8_t* output_validity_;
-  offset_type* output_offsets_;
-  uint8_t* output_values_;
-
- public:
-  ReadWriteValueImpl(const ArraySpan& input_values_array,
-                     ArrayData* output_values_array_data)
-      : input_validity_(has_validity_buffer ? input_values_array.buffers[0].data
-                                            : NULLPTR),
-        input_offsets_(input_values_array.template GetValues<offset_type>(1, 0)),
-        input_values_(input_values_array.buffers[2].data),
-        output_validity_((has_validity_buffer && output_values_array_data)
-                             ? output_values_array_data->buffers[0]->mutable_data()
-                             : NULLPTR),
-        output_offsets_(
-            output_values_array_data
-                ? output_values_array_data->template GetMutableValues<offset_type>(1, 0)
-                : NULLPTR),
-        output_values_(output_values_array_data
-                           ? output_values_array_data->buffers[2]->mutable_data()
-                           : NULLPTR) {}
-
-  [[nodiscard]] bool ReadValue(ValueRepr* out, int64_t read_offset) const {
-    bool valid = true;
-    if constexpr (has_validity_buffer) {
-      valid = bit_util::GetBit(input_validity_, read_offset);
-    }
-    if (valid) {
-      const offset_type offset0 = input_offsets_[read_offset];
-      const offset_type offset1 = input_offsets_[read_offset + 1];
-      *out = std::string_view(reinterpret_cast<const char*>(input_values_ + offset0),
-                              offset1 - offset0);
-    }
-    return valid;
-  }
-
-  /// \brief Ensure padding is zeroed in validity bitmap.
-  void ZeroValidityPadding(int64_t length) const {
-    DCHECK(output_values_);
-    if constexpr (has_validity_buffer) {
-      DCHECK(output_validity_);
-      const int64_t validity_buffer_size = bit_util::BytesForBits(length);
-      output_validity_[validity_buffer_size - 1] = 0;
-    }
-  }
-
-  void WriteValue(int64_t write_offset, bool valid, ValueRepr value) const {
-    if constexpr (has_validity_buffer) {
-      bit_util::SetBitTo(output_validity_, write_offset, valid);
-    }
-    const offset_type offset0 = output_offsets_[write_offset];
-    const offset_type offset1 =
-        offset0 + (valid ? static_cast<offset_type>(value.size()) : 0);
-    output_offsets_[write_offset + 1] = offset1;
-    if (valid) {
-      memcpy(output_values_ + offset0, value.data(), value.size());
-    }
-  }
-
-  void WriteRun(int64_t write_offset, int64_t run_length, bool valid,
-                ValueRepr value) const {
-    if constexpr (has_validity_buffer) {
-      bit_util::SetBitsTo(output_validity_, write_offset, run_length, valid);
-    }
-    if (valid) {
-      int64_t i = write_offset;
-      offset_type offset = output_offsets_[i];
-      while (i < write_offset + run_length) {
-        memcpy(output_values_ + offset, value.data(), value.size());
-        offset += static_cast<offset_type>(value.size());
-        i += 1;
-        output_offsets_[i] = offset;
-      }
-    } else {
-      offset_type offset = output_offsets_[write_offset];
-      offset_type* begin = output_offsets_ + write_offset + 1;
-      std::fill(begin, begin + run_length, offset);
-    }
-  }
-
-  bool Compare(ValueRepr lhs, ValueRepr rhs) const { return lhs == rhs; }
-};
 
 Result<std::shared_ptr<Buffer>> AllocateValuesBuffer(int64_t length, const DataType& type,
                                                      MemoryPool* pool,
@@ -319,7 +57,7 @@ class RunEndEncodingLoop {
   using RunEndCType = typename RunEndType::c_type;
 
  private:
-  using ReadWriteValue = ReadWriteValueImpl<ValueType, has_validity_buffer>;
+  using ReadWriteValue = ree_util::ReadWriteValue<ValueType, has_validity_buffer>;
   using ValueRepr = typename ReadWriteValue::ValueRepr;
 
  private:
@@ -645,7 +383,7 @@ class RunEndDecodingLoop {
   using RunEndCType = typename RunEndType::c_type;
 
  private:
-  using ReadWriteValue = ReadWriteValueImpl<ValueType, has_validity_buffer>;
+  using ReadWriteValue = ree_util::ReadWriteValue<ValueType, has_validity_buffer>;
   using ValueRepr = typename ReadWriteValue::ValueRepr;
 
   const ArraySpan& input_array_;
@@ -660,19 +398,20 @@ class RunEndDecodingLoop {
 
  public:
   RunEndDecodingLoop(const ArraySpan& input_array, ArrayData* output_array_data)
-      : RunEndDecodingLoop(input_array, ree_util::ValuesArray(input_array),
+      : RunEndDecodingLoop(input_array, arrow::ree_util::ValuesArray(input_array),
                            output_array_data) {}
 
   /// \brief For variable-length types, calculate the total length of the data
   /// buffer needed to store the expanded values.
   int64_t CalculateOutputDataBufferSize() const {
-    auto& input_array_values = ree_util::ValuesArray(input_array_);
+    auto& input_array_values = arrow::ree_util::ValuesArray(input_array_);
     DCHECK_EQ(input_array_values.type->id(), ValueType::type_id);
     if constexpr (is_base_binary_like(ValueType::type_id)) {
       using offset_type = typename ValueType::offset_type;
       int64_t data_buffer_size = 0;
 
-      const ree_util::RunEndEncodedArraySpan<RunEndCType> ree_array_span(input_array_);
+      const arrow::ree_util::RunEndEncodedArraySpan<RunEndCType> ree_array_span(
+          input_array_);
       const auto* offsets_buffer =
           input_array_values.template GetValues<offset_type>(1, 0);
       auto it = ree_array_span.begin();
@@ -695,7 +434,8 @@ class RunEndDecodingLoop {
   ARROW_NOINLINE int64_t ExpandAllRuns() {
     read_write_value_.ZeroValidityPadding(input_array_.length);
 
-    const ree_util::RunEndEncodedArraySpan<RunEndCType> ree_array_span(input_array_);
+    const arrow::ree_util::RunEndEncodedArraySpan<RunEndCType> ree_array_span(
+        input_array_);
     int64_t write_offset = 0;
     int64_t output_valid_count = 0;
     for (auto it = ree_array_span.begin(); !it.is_end(ree_array_span); ++it) {
@@ -774,7 +514,8 @@ struct RunEndDecodeExec {
     if constexpr (ValueType::type_id == Type::NA) {
       return RunEndDecodeNullREEArray(ctx, input_array, result);
     } else {
-      const bool has_validity_buffer = ree_util::ValuesArray(input_array).MayHaveNulls();
+      const bool has_validity_buffer =
+          arrow::ree_util::ValuesArray(input_array).MayHaveNulls();
       if (has_validity_buffer) {
         return RunEndDecodeImpl<RunEndType, ValueType, true>(ctx, input_array, result)
             .Exec();
